# swagger.yaml
openapi: 3.0.0
info:
  title: E-Commerce Platform API
  version: 1.0.0
  description: |
    REST API for an E-Commerce platform built with Node.js, Express, and PostgreSQL.

    **Role IDs in JWT:**
    | role_id | Role     | Access                                    |
    |---------|----------|-------------------------------------------|
    | 1       | Admin    | Full system access                        |
    | 2       | Customer | Browse, cart, orders, wishlist            |
    | 3       | Vendor   | Manage own products                       |
    | 4       | Courier  | View and update assigned deliveries       |

    **Transaction endpoints:**
    - `POST /orders` — BEGIN/COMMIT/ROLLBACK, stock deduction via DB trigger
    - `POST /admin/assign-courier` — BEGIN/COMMIT/ROLLBACK, row-level locks
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server

# ─── Global security — override with `security: []` on public routes ──────────
security:
  - bearerAuth: []

# ─── Reusable components ───────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /auth/login. Include as: Authorization Bearer token."
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Resource not found"
        detail:
          type: string
          example: "Additional context about the failure"

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    Category:
      type: object
      properties:
        id:            { type: integer, example: 1 }
        category_name: { type: string,  example: "Electronics" }
        parent_cat_id: { type: integer, nullable: true, example: null }
        parent_name:   { type: string,  nullable: true, example: null }
        status:        { type: string,  enum: [active, inactive], example: "active" }
        created_at:    { type: string,  format: date-time }

    # ── Product ────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:            { type: integer, example: 1 }
        product_name:  { type: string,  example: "Smartphone X" }
        description:   { type: string,  example: "Latest state-of-the-art smartphone." }
        category_name: { type: string,  example: "Electronics" }
        status:        { type: string,  enum: [active, inactive], example: "active" }
        variant_count: { type: integer, example: 3 }
        created_at:    { type: string,  format: date-time }
...
    CartItem:
      type: object
      properties:
        cart_item_id:        { type: integer, example: 1 }
        product_variant_id:  { type: integer, example: 3 }
        product_name:        { type: string,  example: "Smartphone X" }
        variant_name:        { type: string,  example: "128GB Black" }
        price:               { type: number,  example: 799.99 }
        stock_quantity:      { type: integer, example: 50 }
        quantity:            { type: integer, example: 2 }
        line_total:          { type: number,  example: 1599.98 }
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
        total:
          type: string
          example: "1599.98"

    # ── Order ──────────────────────────────────────────────────────────────────
...
    OrderSummary:
      type: object
      properties:
        id:              { type: integer, example: 1 }
        order_number:    { type: string,  example: "ORD-1716000000000-427" }
        total_amount:    { type: string,  example: "1599.98" }
        discount_amount: { type: string,  example: "160.00" }
        shipping_amount: { type: string,  example: "5.99" }
        net_amount:      { type: string,  example: "1445.97" }
        status:          { type: string,  enum: [placed, processing, shipping, delivered] }
        payment_status:  { type: string,  enum: [paid, "not paid"] }
        payment_type:    { type: string,  enum: [netbanking, upi, cod] }
        created_at:      { type: string,  format: date-time }
...
    Offer:
      type: object
      properties:
        id:              { type: integer, example: 1 }
        coupon_code:     { type: string,  example: "SUMMER10" }
        discount_type:   { type: string,  enum: [fixed, rate] }
        discount_value:  { type: number,  example: 10.00 }
        start_date:      { type: string,  format: date, example: "2026-06-01" }
        end_date:        { type: string,  format: date, example: "2026-08-31" }
        status:          { type: string,  enum: [active, inactive] }

    # ── Delivery ───────────────────────────────────────────────────────────────
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List all active categories (public)
      security: []
      responses:
        '200':
          description: Array of active categories with parent name resolved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }

    post:
      tags: [Categories]
      # FIX: was "Admin/Vendor" — categories are platform-wide taxonomy, Admin only
      summary: Create a category (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_name]
              properties:
                category_name: { type: string,  example: "Electronics" }
                parent_cat_id: { type: integer, nullable: true, example: null }
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404':
          description: Parent category not found or inactive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /products:
    get:
      tags: [Products]
      summary: List active products, paginated (public)
      security: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
          description: Page number
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
          description: Results per page (max 100)
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
                  pagination:
                    type: object
                    properties:
                      page:        { type: integer, example: 1 }
                      limit:       { type: integer, example: 20 }
                      total:       { type: integer, example: 84 }
                      total_pages: { type: integer, example: 5 }

    post:
      tags: [Products]
      summary: Create a product (Admin or Vendor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, product_name]
              properties:
                category_id:  { type: integer, example: 1 }
                product_name: { type: string,  example: "Smartphone X" }
                description:  { type: string,  example: "Latest state-of-the-art smartphone." }
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Category not found or inactive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /cart:
    # FIX: was POST /cart/add — REST convention is POST to the resource root
    post:
      tags: [Cart]
      summary: Add an item to the cart, or increment quantity if already present (Customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_variant_id]
              properties:
                product_variant_id: { type: integer, example: 3 }
                quantity:           { type: integer, example: 1, minimum: 1 }
      responses:
        '200':
          description: Item added or quantity updated
        '400':
          description: Quantity exceeds available stock (includes already-in-cart amount)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Product variant not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Cart]
      summary: View current cart with line totals and cart total (Customer only)
      responses:
        '200':
          description: Cart contents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse' }
...
  /wishlist:
    post:
      tags: [Wishlist]
      summary: Add a product variant to wishlist (Customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_variant_id]
              properties:
                product_variant_id: { type: integer, example: 3 }
      responses:
        '201':
          description: Item added to wishlist
        '404':
          description: Product variant not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Item is already in the wishlist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Wishlist]
      summary: Get the user's wishlist (Customer only)
      responses:
        '200':
          description: Wishlist items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        wishlist_id:        { type: integer }
                        product_variant_id: { type: integer }
                        product_name:       { type: string }
                        variant_name:       { type: string }
                        price:              { type: number }
                        stock_quantity:     { type: integer }
...
  /orders:
    post:
      tags: [Orders]
      summary: Place an order from cart (Customer only, transactional)
      description: |
        Runs inside a PostgreSQL transaction.
        Steps: load cart → stock pre-check → validate coupon → INSERT Order →
        INSERT Order_items (DB trigger `trg_reduce_stock` deducts stock per row) →
        DELETE Cart_items → COMMIT.
        Any failure (including trigger stock exception) triggers a full ROLLBACK.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_type]
              properties:
                payment_type:
                  type: string
                  enum: [netbanking, upi, cod]
                  example: "upi"
                shipping_amount:
                  type: number
                  example: 5.99
                  default: 0
                # FIX: was `discount_amount` (client-controlled — security risk).
                # Discount is now derived server-side from the coupon code only.
                coupon_code:
                  type: string
                  example: "SUMMER10"
                  description: Optional coupon code. Validated against the Offers table.
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:         { type: string, example: "Order created successfully" }
                  order_id:        { type: integer, example: 7 }
                  order_number:    { type: string,  example: "ORD-1716000000000-427" }
                  total_amount:    { type: string,  example: "1599.98" }
                  discount_amount: { type: string,  example: "160.00" }
                  shipping_amount: { type: string,  example: "5.99" }
                  net_amount:      { type: string,  example: "1445.97" }
        '400':
          description: Cart empty, invalid coupon, or invalid payment type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Insufficient stock — transaction rolled back
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Orders]
      summary: List orders (Admin sees all; Customer sees own)
      responses:
        '200':
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrderSummary' }
...
  /offers:
    get:
      tags: [Offers]
      summary: List currently active offers (public)
      security: []
      description: Only offers with `status=active` and within date range are returned.
      responses:
        '200':
          description: Active offers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Offer' }

    # FIX: was "Admin/Vendor" — platform coupons are Admin only
    post:
      tags: [Offers]
      summary: Create a discount offer (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coupon_code, discount_type, discount_value, start_date, end_date]
              properties:
                coupon_code:    { type: string, example: "SUMMER10" }
                discount_type:  { type: string, enum: [fixed, rate], example: "rate" }
                discount_value: { type: number, example: 10.00 }
                start_date:     { type: string, format: date, example: "2026-06-01" }
                end_date:       { type: string, format: date, example: "2026-08-31" }
      responses:
        '201':
          description: Offer created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Offer' }
        '400':
          description: Invalid discount type, value, or date range
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Coupon code already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # FIX: This endpoint was missing from the original swagger entirely
...
  /courier/orders:
    get:
      tags: [Courier]
      summary: Get deliveries assigned to the logged-in courier (Courier only)
      responses:
        '200':
          description: List of assigned delivery records with order and address info
        '404':
          description: Courier profile not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...